+++ src/form.rs
@@ -156,8 +156,8 @@ where
     type SerializeStruct = Self;
     type SerializeStructVariant = Self;
 
-    fn serialize_bool(self, _v: bool) -> Result<Self::Ok, Self::Error> {
-        Err(QuerylizerError::UnsupportedValue)
+    fn serialize_bool(self, v: bool) -> Result<Self::Ok, Self::Error> {
+        self.serialize_str(if v { "true" } else { "false" })
     }
 
     fn serialize_i8(self, v: i8) -> Result<Self::Ok, Self::Error> {
@@ -564,8 +564,12 @@ mod tests {
     #[test]
     fn test_bool() -> Result<(), QuerylizerError> {
         assert_eq!(
-            Form::to_string("color", &true, false, passthrough),
-            Err(QuerylizerError::UnsupportedValue)
+            Form::to_string("color", &true, false, passthrough)?,
+            "color=true"
+        );
+        assert_eq!(
+            Form::to_string("color", &false, false, passthrough)?,
+            "color=false"
         );
         Ok(())
     }
@@ -729,10 +733,7 @@ mod tests {
 
     #[test]
     fn test_unit() -> Result<(), QuerylizerError> {
-        assert_eq!(
-            Form::to_string("color", &(), false, passthrough)?,
-            "color="
-        );
+        assert_eq!(Form::to_string("color", &(), false, passthrough)?, "color=");
         Ok(())
     }