+++ CHANGES.md
@@ -1,8 +1,9 @@
 # Querylizer
 
-### Next
+### v0.1.3 (2022-03-03)
 
 - Support `boolean` serialization.
+- Always encode `+` in query to avoid historical ambiguity.
 
 ### v0.1.2 (2022-03-03)